Delete db popup created and map screen allows non db locations

+++ hadam_A4/app/src/main/java/com/csci448/hadam/hadam_a4/presentation/settings/SettingsScreen.kt
@@ -7,10 +7,13 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
+import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
@@ -27,9 +30,11 @@ fun SettingsScreen(
 ) {
     val saveEnabled = historyViewModel.saveLocationsEnabled.collectAsStateWithLifecycle().value
     Column() {
-        Row(modifier = Modifier
-            .fillMaxWidth()
-            .padding(10.dp)) {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(10.dp)
+        ) {
             Text(
                 text = "Save Locations to Database",
                 modifier = Modifier.padding(top = 10.dp, end = 40.dp),
@@ -40,19 +45,47 @@ fun SettingsScreen(
             })
         }
 
-
+        val showDialog = remember { mutableStateOf(false) }
         Button(
             onClick = {
-                historyViewModel.deleteHistory()
-                Toast.makeText(
-                    context,
-                    "History deleted",
-                    Toast.LENGTH_LONG
-                ).show()
+                showDialog.value = true
+
             },
-            modifier = Modifier.padding(start = 20.dp, end = 20.dp).fillMaxWidth()
+            modifier = Modifier
+                .padding(start = 20.dp, end = 20.dp)
+                .fillMaxWidth()
         ) {
             Text(text = "Delete Database")
+        }
+        if (showDialog.value) {
+            AlertDialog(
+                onDismissRequest = { showDialog.value = false },
+                dismissButton = {
+                    Button(
+                        onClick = {
+                            showDialog.value = false
+                        },
+                        content = { Text(text = "No") }
+                    )
+                },
+                title = { Text(text = "Delete Database") },
+                text = { Text(text = "Are you sure you want to delete the database?") },
+                confirmButton = {
+                    Button(
+                        onClick = {
+                            showDialog.value = false
+
+                            historyViewModel.deleteHistory()
+                            Toast.makeText(
+                                context,
+                                "History deleted",
+                                Toast.LENGTH_LONG
+                            ).show()
+                        },
+                        content = { Text(text = "Yes") }
+                    )
+                }
+            )
         }
     }
 }
